pyproject.toml
[tool.black]
line-length = 75

[tool.flake8]
max-line-length = 75
exclude = ["venv/*", ".git/*", "__pycache__/*"]


[tool.poetry]
name = "youtube-audio-ripper"
version = "0.1.0"
description = "Flask API Server to strip audio files from YouTube URL"
authors = ["Gray Man"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
flasgger = "^0.9.7.1"
moviepy = "^1.0.3"
pytubefix = "^8.1.1"
boto3 = "^1.35.44"
sqlalchemy = "^2.0.36"
pymysql = "^1.1.1"
flask-sqlalchemy = "^3.1.1"
flask-cors = "^5.0.0"

[tool.poetry.dev-dependencies]
isort = "^5.13.2"
black = "^24.10.0"
flake8 = "^7.1.1"
pytest-flask = "^1.3.0"
pytest-cov = "^5.0.0"
pytest = "^8.3.3"

[tool.poetry.group.dev.dependencies]
commitizen = "^3.29.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"